refactor(tvix/eval): shrink operations in bytecode to u8 representation
WIP! This commit message is not done yet, just notes so far.

Current status:

- slightly faster, but very much (consistently so, but within what between 1-2
  runs could be considered margin of error)
- but: not all optimisations that we did before are reenabled yet

I'm actually undecided on whether to work on reenabling them, or take the
current state and get rid of the optimisation passes for now (esp. unthunking),
and putting them back in the known optimisation potential list.

Having less code in eval right now would be good, because the next changes I
want to do are more drastic.

Change-Id: Id9b3a4254623a45de03069dbdb70b8349e976743