feat(tvix/glue): add TvixStoreIO::realize_store_path
Add a new `realize_store_path` method to TvixStoreIO, which only ensures
that the given path (which must be a store path or a subpath of a store
path) has been built.

Under the hood, this factors out a `realize_store_path` method from
`store_path_to_node`, and uses that.

Change-Id: Ide53febdef054ac9f09fa75d6e0c227383e62dbc