WIP/BLOCKED: feat(tvix/utils): Add mkCrate2nixCheck
This adds a function which can be used across the monorepo to create a
derivation that checks whether the Cargo.nix file is up-to-date. This is
meant to be used for crates that depend on //tvix.

Currently blocked because of a bug in 2.3,
waiting for CI to be upgraded to a newer Nix version.

Change-Id: Idc6c30aded4384f425e7b07cea6a26e4ee9f5db5