feat(ops/pipelines): support buildkite retries
cl/12228 did enable automatic retries for some flaky tests, which
generally did work, as can be seen in
https://buildkite.com/tvl/depot/builds/35893

However, ":duck:" still reports as failing, because we check the number
of steps to be nonzero, which is not the case if retries have happened.

Instead we can check for the total build to be not in the "FAILING"
state.

Change-Id: Ib9d27587c8a8ba7970850812c4302fecdc4482e7