feat(nix-compat/narinfo/signature): add owned Signature version
Rename our existing Signature to SignatureRef, and add an owned version
of it, using SmolStr for the signature name.
Implement Display, Serialize and Deserialize too.

Have as_ref() and to_owned() helpers to allow conversion back and forth.

This is in preparation for a stronger typed `PathInfo` struct, where we
want to have full ownership.

Change-Id: I3f75691498c6bda9cd072d2d9dac83c4f6c57287