chore(tvix/eval): rename Evaluation{Builder,} → Evaluator{Builder,}
Currently, the evaluation machinery is set up for one-shot evaluations.

In context where we want to write a program that drive the evaluation
process partially, e.g. force certain thunks based on custom traversal
code, e.g. nix-eval-jobs. It is necessary to keep the VM and not
re-instantiate all the time.

We rename to signal to users that this entity will now be set up so that
continuous evaluation requests can be sent.

Change-Id: I1724fa1d25efcdc68d0e6319e3183642e45eee8d
Signed-off-by: Ryan Lahfa <tvl@lahfa.xyz>