feat(tvix/eval): VM is now reusable for subsequent executions
Previously, the VM was consuming itself at the `execute` call.

With this, the VM is just mutably borrowed at `execute` call and the
VM's warnings are consumed in the result of the evaluation.

Tests needs to be written to verify we can indeed perform subsequent
executions without internal state issues.

Change-Id: Iaeeab7658f9908734f7183526c864b43537fb1a7
Signed-off-by: Ryan Lahfa <tvl@lahfa.xyz>