DO NOT MERGE: debug NarCalculationService
We currently don't successfully use the NAR calculation service provided
by a GRPCPathInfoService.

This means, a `tvix-store import` using the defaults will fall back
to `SimpleRenderer`, which will calculate the NAR hash and size by
downloading the uploaded blobs (and blobs it didn't need to upload)
locally. This is of course silly.

That's where

```rust
impl PathInfoService {
  fn nar_calculation_service(&self)
    -> Option<Box<dyn NarCalculationService>>;
}
```

should come into play, and needs to be implemented for
`GRPCPathInfoService`.

It however still doesn't seem to work work - we still see the
"Falling back to simple renderer…" log line being emitted for both
`tvix-store import` and `tvix-store daemon`, it should only be logged
for the daemon.

Change-Id: If2c3fe6584e9093cba322d2360f355a3923904ae