fix(tvix/shell): fix shell on MacOS
runc is not available on MacOS, we only want to include it in the shell if on
Linux.

Point TVIX_BUILD_SANDBOX_SHELL to /bin/sh if not on Linux.
While we cannot use the OCI/runc executor on MacOS, other implementations might
make use of this env var, so it's good to keep setting it.

Also update the other occurence of it, in tvix/utils.nix.

Change-Id: If7b47e1bb7c41bbde84c93016713754a252c4355