feat(nix-compat-derive): Add nix serialization support.
As part of its operation, the nix-daemon needs to both read and write
structured data as it communicates with its clients.

nix-compat-derive already has support for deserializing values off the
wire. This change adds serialization support and automatic deriving
based on `NixSerialize` trait.

This cl does not allow automatic deriving of `NixSerialize` for enums
mainly due to the fact that conditional serialization like the one
`NixDeserialize` supports is not possible. That's because by the time
the version is known, the value to serialize has already been created,
so it's up to the application logic to determine which enum to
construct.

When(and most importantly *if*) enum support becomes needed, it can be
added at that point.

Change-Id: If171fb182b936da88287133deb8bbf01259e03f7